Refactor: 리뷰 개수는 Long으로, 별점 평균은 Double로 타입 설정

+++ src/main/java/kopis/k_backend/performance/domain/Performance.java
@@ -51,7 +51,7 @@ public class Performance {
 
     private String poster;
 
-    private Long ratingAverage = 0L;
+    private Double ratingAverage = 0D;
 
     private Long reviewCount = 0L;
